webscrape/requirements.txt
ipykernel
nbformat
black
pandas
plotly
requests
numpy
seaborn
watermark
matplotlib
beautifulsoup4
kaleido
hydra-core
lxml
python-json-logger